# `AsyncLogger`: enqueue logging operation instead of a log message