Original upstream PR 👉🏽 https://github.com/valskalla/odin/pull/461

There are some differences between this PR and original one, the most notable one, not using a `Semaphore` around `drain`. AFAICS, `drain` is only called internally, and spaced by a time window. I don't see the reason to use permits, given that only one fiber (the background loop) will be calling it.

The part that concerns me is the benchmark. Here we're trading performance for correctness, which is a good point, but I'd like to hear some users' thoughts about it.

Now the results, following the same approach as upstream PR:

## 1) AsyncLoggerBenchmark
### Before
```
[info] Benchmark                             Mode  Cnt     Score     Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  1635.576 ± 145.618  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  1713.306 ± 209.162  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  2002.529 ± 939.309  ns/op
```

### After
```
[info] Benchmark                             Mode  Cnt     Score      Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  1803.819 ±  159.555  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  1895.908 ±  412.442  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  3095.106 ± 1259.061  ns/op
```

## 2) Measure the cost of the `drain` operation
### Before
```
[info] Benchmark                             Mode  Cnt     Score     Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  1607.056 ± 266.458  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  1587.006 ± 133.531  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  6903.186 ± 186.564  ns/op
```

### After
```
[info] Benchmark                             Mode  Cnt     Score     Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  4044.233 ± 126.506  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  4215.031 ± 111.849  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  9383.689 ± 216.399  ns/op
```

cc @iRevive @ybasket