@iRevive 2024-08-10
Side note: the original `AsyncLoggerBenchmark ` benchmark is somewhat rigged.

The benchmark calls the `logger.info`:
```scala
@Benchmark
@OperationsPerInvocation(1000)
def msg(): Unit = (1 to 1000).toList.traverse_(_ => asyncLogger.info(message)).unsafeRunSync()
```
 
 `logger.info` calls `submit` under the hood, which `AsyncLogger` defines as:
 ```scala
 def submit(msg: LoggerMessage): F[Unit] = queue.tryOffer(msg).void
 ```
 
 The benchmark measures how fast the `cats.effect.Queue#tryOffer` is implemented. 
 
 Perhaps the `drain` benchmark could be incorporated to measure the actual throughput.