@iRevive 2024-08-10
Regarding the throughput: 

Let's say the `drain` function is invoked every 10 milliseconds. It takes `4044` nanoseconds to evaluate one 'log' invocation. 
So, the max number of 'log' evaluations the system can process within the 10 ms window is `2472`.

If the application emits 1k logs per second, the background logging will fall behind quickly.

-----

In my opinion, it makes sense to implement a checker to track the processing lag. If the lag is greater than the window, we can warn users.

For example:
```scala
val drainLoop = {
  val step = 
    for {
      time <- drain.timed._1F
      _    <- logger.error("The async logger falls behind by [${time - window}]").whenA(time > window)
    } yield ()
  
  step.delayBy(window).foreverM
}
```