@iRevive 2024-08-10
> > In my opinion, it makes sense to implement a checker to track the processing lag. If the lag is greater than the window, we can warn users.
> > For example:
> > ```scala
> > val drainLoop = {
> >   val step = 
> >     for {
> >       time <- drain.timed._1F
> >       _    <- logger.error("The async logger falls behind by [${time - window}]").whenA(time > window)
> >     } yield ()
> >   
> >   step.delayBy(window).foreverM
> > }
> > ```
> 
> To me, this is overengineering for an edge case. I'd really like to explore other options. Even not allowing changing the level of an `AsyncLogger`.

Yeah. I believe 99% of the users will never hit this case 🙂