@iRevive 2024-08-10
> Even not allowing changing the level of an AsyncLogger.

It's necessary for conditional logging scenarios when you need to change the log level on-fly. 
Also, `withMinimalLevel` comes from the very-base `Logger` interface. 

So it would be nice to have it working on all loggers.