@aartigao 2024-08-10
> Also, `withMinimalLevel` comes from the very-base `Logger` interface.

Yep. IMO, this is what it makes it difficult. Some loggers like `ConsoleLogger` can be purely constructed, but some others, like `AsyncLogger` or `FileLogger` can be constructed via `Resource` only. This is incompatible with `withMinimalLevel`, because they're basically playing different leagues. `FileLogger` can solve it easily, because the constructor was used only to create the file, so changing the level is seamless. But `AsyncLogger` has a background fiber, and this is where things get tricky.

The proposed solution invalidates the batch processing in the inner logger. This, combined with inner logger of `FileLogger` gives fatal results, because each log message is flushed to the file. If we can't workaround this, we're f****d 🙁