@aartigao 2024-08-10
Things get better 👇🏽

```
[info] Benchmark                             Mode  Cnt     Score      Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  2599.208 ±  174.239  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  2698.314 ±  179.909  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  8701.166 ± 1059.788  ns/op
```

But still far from good!