@aartigao 2024-08-10
LOL, I think I've even improved the original `drain` benchmarks 😬 👇🏽 

```
[info] Benchmark                             Mode  Cnt     Score     Error  Units
[info] AsyncLoggerBenchmark.msg              avgt   25  1201.403 ± 120.967  ns/op
[info] AsyncLoggerBenchmark.msgAndCtx        avgt   25  1285.561 ±  83.073  ns/op
[info] AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  6739.552 ± 352.120  ns/op
```

I'll fix the tests and update the changes for review.